# dms/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DMS API - PT Askara Internal",
    "description": "Document Management System API for PT Askara Internal. Manages documents, workflows, templates, and organizational structure.",
    "version": "1.0.0",
    "contact": {
      "name": "PT Askara Internal",
      "email": "admin@askara.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Login successful", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Logout successful"}}
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh token",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Token refreshed"}}
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Current user info"}}
      }
    },
    "/auth/password": {
      "put": {
        "tags": ["Auth"],
        "summary": "Change password",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"current_password": {"type": "string"}, "new_password": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Password changed"}}
      }
    },
    "/companies": {
      "get": {"tags": ["Companies"], "summary": "List companies", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/PageParam"}, {"$ref": "#/components/parameters/PerPageParam"}], "responses": {"200": {"description": "List of companies"}}},
      "post": {"tags": ["Companies"], "summary": "Create company", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CompanyInput"}}}}, "responses": {"201": {"description": "Company created"}}}
    },
    "/companies/{id}": {
      "get": {"tags": ["Companies"], "summary": "Get company", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Company details"}}},
      "put": {"tags": ["Companies"], "summary": "Update company", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Company updated"}}},
      "delete": {"tags": ["Companies"], "summary": "Delete company", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Company deleted"}}}
    },
    "/offices": {
      "get": {"tags": ["Offices"], "summary": "List offices", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of offices"}}},
      "post": {"tags": ["Offices"], "summary": "Create office", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OfficeInput"}}}}, "responses": {"201": {"description": "Office created"}}}
    },
    "/offices/{id}": {
      "get": {"tags": ["Offices"], "summary": "Get office", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Office details"}}},
      "put": {"tags": ["Offices"], "summary": "Update office", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Office updated"}}},
      "delete": {"tags": ["Offices"], "summary": "Delete office", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Office deleted"}}}
    },
    "/offices/{id}/departments": {
      "get": {"tags": ["Offices"], "summary": "List departments in office", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Departments"}}}
    },
    "/departments": {
      "get": {"tags": ["Departments"], "summary": "List departments", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of departments"}}},
      "post": {"tags": ["Departments"], "summary": "Create department", "security": [{"bearerAuth": []}], "responses": {"201": {"description": "Department created"}}}
    },
    "/departments/{id}": {
      "get": {"tags": ["Departments"], "summary": "Get department", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Department details"}}},
      "put": {"tags": ["Departments"], "summary": "Update department", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Department updated"}}},
      "delete": {"tags": ["Departments"], "summary": "Delete department", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Department deleted"}}}
    },
    "/departments/{id}/sections": {
      "get": {"tags": ["Departments"], "summary": "List sections in department", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Sections"}}}
    },
    "/sections": {
      "get": {"tags": ["Sections"], "summary": "List sections", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of sections"}}},
      "post": {"tags": ["Sections"], "summary": "Create section", "security": [{"bearerAuth": []}], "responses": {"201": {"description": "Section created"}}}
    },
    "/sections/{id}": {
      "get": {"tags": ["Sections"], "summary": "Get section", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Section details"}}},
      "put": {"tags": ["Sections"], "summary": "Update section", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Section updated"}}},
      "delete": {"tags": ["Sections"], "summary": "Delete section", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Section deleted"}}}
    },
    "/positions": {
      "get": {"tags": ["Positions"], "summary": "List positions", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of positions"}}},
      "post": {"tags": ["Positions"], "summary": "Create position", "security": [{"bearerAuth": []}], "responses": {"201": {"description": "Position created"}}}
    },
    "/positions/{id}": {
      "get": {"tags": ["Positions"], "summary": "Get position", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Position details"}}},
      "put": {"tags": ["Positions"], "summary": "Update position", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Position updated"}}},
      "delete": {"tags": ["Positions"], "summary": "Delete position", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Position deleted"}}}
    },
    "/users": {
      "get": {"tags": ["Users"], "summary": "List users", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/PageParam"}, {"$ref": "#/components/parameters/PerPageParam"}, {"name": "search", "in": "query", "schema": {"type": "string"}}], "responses": {"200": {"description": "List of users"}}},
      "post": {"tags": ["Users"], "summary": "Create user", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserInput"}}}}, "responses": {"201": {"description": "User created"}}}
    },
    "/users/{id}": {
      "get": {"tags": ["Users"], "summary": "Get user", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "User details"}}},
      "put": {"tags": ["Users"], "summary": "Update user", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "User updated"}}},
      "delete": {"tags": ["Users"], "summary": "Delete user", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "User deleted"}}}
    },
    "/users/{id}/signature": {
      "post": {"tags": ["Users"], "summary": "Upload signature", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "requestBody": {"content": {"multipart/form-data": {"schema": {"type": "object", "properties": {"signature": {"type": "string", "format": "binary"}}}}}}, "responses": {"200": {"description": "Signature uploaded"}}},
      "delete": {"tags": ["Users"], "summary": "Delete signature", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Signature deleted"}}}
    },
    "/users/{id}/avatar": {
      "post": {"tags": ["Users"], "summary": "Upload avatar", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "requestBody": {"content": {"multipart/form-data": {"schema": {"type": "object", "properties": {"avatar": {"type": "string", "format": "binary"}}}}}}, "responses": {"200": {"description": "Avatar uploaded"}}}
    },
    "/users/{id}/roles": {
      "put": {"tags": ["Users"], "summary": "Assign roles", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"role_ids": {"type": "array", "items": {"type": "string", "format": "uuid"}}}}}}}, "responses": {"200": {"description": "Roles assigned"}}}
    },
    "/roles": {
      "get": {"tags": ["Roles"], "summary": "List roles", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of roles"}}}
    },
    "/roles/{id}": {
      "get": {"tags": ["Roles"], "summary": "Get role with permissions", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Role details"}}}
    },
    "/permissions": {
      "get": {"tags": ["Roles"], "summary": "List permissions", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of permissions"}}}
    },
    "/document-types": {
      "get": {"tags": ["Document Types"], "summary": "List document types", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of document types"}}},
      "post": {"tags": ["Document Types"], "summary": "Create document type", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DocumentTypeInput"}}}}, "responses": {"201": {"description": "Document type created"}}}
    },
    "/document-types/{id}": {
      "get": {"tags": ["Document Types"], "summary": "Get document type", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Document type details"}}},
      "put": {"tags": ["Document Types"], "summary": "Update document type", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Document type updated"}}}
    },
    "/categories": {
      "get": {"tags": ["Categories"], "summary": "List categories", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of categories"}}},
      "post": {"tags": ["Categories"], "summary": "Create category", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategoryInput"}}}}, "responses": {"201": {"description": "Category created"}}}
    },
    "/categories/{id}": {
      "get": {"tags": ["Categories"], "summary": "Get category", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Category details"}}},
      "put": {"tags": ["Categories"], "summary": "Update category", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Category updated"}}},
      "delete": {"tags": ["Categories"], "summary": "Delete category", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Category deleted"}}}
    },
    "/templates": {
      "get": {"tags": ["Templates"], "summary": "List templates", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/PageParam"}, {"$ref": "#/components/parameters/PerPageParam"}, {"name": "document_type_id", "in": "query", "schema": {"type": "string", "format": "uuid"}}, {"name": "search", "in": "query", "schema": {"type": "string"}}], "responses": {"200": {"description": "List of templates"}}},
      "post": {"tags": ["Templates"], "summary": "Create template", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TemplateInput"}}}}, "responses": {"201": {"description": "Template created"}}}
    },
    "/templates/{id}": {
      "get": {"tags": ["Templates"], "summary": "Get template", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Template details with tags"}}},
      "put": {"tags": ["Templates"], "summary": "Update template", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Template updated"}}},
      "delete": {"tags": ["Templates"], "summary": "Delete template", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "Template deleted"}}}
    },
    "/templates/{id}/upload": {
      "post": {"tags": ["Templates"], "summary": "Upload template DOCX", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "requestBody": {"content": {"multipart/form-data": {"schema": {"type": "object", "properties": {"file": {"type": "string", "format": "binary"}}}}}}, "responses": {"200": {"description": "File uploaded"}}}
    },
    "/templates/{id}/download": {
      "get": {"tags": ["Templates"], "summary": "Download template", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "DOCX file"}}}
    },
    "/templates/{id}/tags": {
      "get": {"tags": ["Templates"], "summary": "List template tags", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "responses": {"200": {"description": "List of tags"}}},
      "post": {"tags": ["Templates"], "summary": "Add tag", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TemplateTagInput"}}}}, "responses": {"201": {"description": "Tag added"}}}
    },
    "/templates/{id}/tags/bulk": {
      "post": {"tags": ["Templates"], "summary": "Bulk upsert tags", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/IdParam"}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"tags": {"type": "array", "items": {"$ref": "#/components/schemas/TemplateTagInput"}}}}}}}, "responses": {"200": {"description": "Tags upserted"}}}
    },
...
    "/numbering/preview": {
      "get": {"tags": ["Numbering"], "summary": "Preview generated number", "security": [{"bearerAuth": []}], "parameters": [{"name": "document_type_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Preview number"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "IdParam": {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "PageParam": {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
      "PerPageParam": {"name": "per_page", "in": "query", "schema": {"type": "integer", "default": 20, "maximum": 100}}
    },
    "schemas": {
      "TokenResponse": {"type": "object", "properties": {"data": {"type": "object", "properties": {"token": {"type": "string"}, "refresh_token": {"type": "string"}, "expires_at": {"type": "string", "format": "date-time"}}}}},
      "CompanyInput": {"type": "object", "required": ["name", "code"], "properties": {"name": {"type": "string"}, "code": {"type": "string"}, "address": {"type": "string"}, "phone": {"type": "string"}, "email": {"type": "string", "format": "email"}}},
      "OfficeInput": {"type": "object", "required": ["name", "code", "type"], "properties": {"company_id": {"type": "string", "format": "uuid"}, "name": {"type": "string"}, "code": {"type": "string"}, "type": {"type": "string", "enum": ["headquarters", "branch", "representative"]}}},
...
    }
  }
}